rename file commodityDetail.js in floder container to commodityDetailMask
